IC-22, IC-26
